requirements.txt
requests
beautifulsoup4
django-dotenv==1.4.1
lxml
